+++ CHANGELOG.md
@@ -3,6 +3,31 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [4.7.7] - 2024-05-10
+### Maintenance
+- PB-33321 Fix local storage loading on extension update
+
+## [4.7.6] - 2024-05-08
+### Maintenance
+- PB-33307 Add debug to capture onInstall reason details
+
+## [4.7.5] - 2024-05-08
+### Maintenance
+- PB-33307 Browser extension version bump to v4.7.5-rc.0
+
+## [4.7.4] - 2024-05-08
+### Maintenance
+- PB-33304 Fix extension update available service
+
+## [4.7.3] - 2024-05-07
+### Maintenance
+- PB-33235 Convert formData file into a json serializable in offscreen
+- PB-33297 Extension update available should store the state if user signed in
+
+## [4.7.1] - 2024-05-02
+### Maintenance
+- PB-33225 MV3 beta rollout
+
 ## [4.7.0] - 2024-04-26
 ### Added
 - PB-32931 As administrator, I see SSO and Directory Sync health checks in Passbolt API Status page
@@ -1586,7 +1611,13 @@ self registration settings option in the left-side bar
 - AP: User with plugin installed
 - LU: Logged in user
 
-[Unreleased]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.0...HEAD
+[Unreleased]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.7...HEAD
+[4.7.7]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.6...v4.7.7
+[4.7.6]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.5...v4.7.6
+[4.7.5]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.4...v4.7.5
+[4.7.4]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.3...v4.7.4
+[4.7.3]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.1...v4.7.3
+[4.7.1]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.7.0...v4.7.1
 [4.7.0]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.6.2...v4.7.0
 [4.6.2]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.6.0...v4.6.2
 [4.6.0]: https://github.com/passbolt/passbolt_browser_extension/compare/v4.5.2...v4.6.0